Refactor doctor verification request model and fields
Updates the doctor verification request system with improved data modeling:

- Renames RequestedAtUtc to CreatedAtUtc for clarity
- Changes RejectionNotes to RejectionNote for consistency
- Modifies Id fields from int to long for scalability
- Adds new entities for doctor languages and working hours
- Introduces timezone and nationality fields
- Adds configurations for new relationships and constraints
- Implements grouping extension method for working hours

These changes provide better support for international doctors and working hour management while maintaining data integrity through proper configurations

Enhance doctor verification with international support

Improves doctor verification system to better handle international use cases:

- Changes ID fields from int to long for better scalability
- Adds support for timezone and nationality tracking
- Introduces doctor languages and working hours entities
- Standardizes field naming (CreatedAtUtc, RejectionNote)
- Configures proper relationships and constraints for new entities

This refactoring enables better management of international doctors while ensuring data integrity and scalability.

+++ Seravian/Services/DoctorsVerificationRequestsAttachmentFilesAccessLockingManager.cs
@@ -3,15 +3,15 @@ using Nito.AsyncEx;
 
 public class DoctorsVerificationRequestsAttachmentFilesAccessLockingManager
 {
-    private readonly ConcurrentDictionary<int, AsyncReaderWriterLock> _locks = [];
+    private readonly ConcurrentDictionary<long, AsyncReaderWriterLock> _locks = [];
 
-    public async Task<IDisposable> EnterReadAsync(int requestId)
+    public async Task<IDisposable> EnterReadAsync(long requestId)
     {
         var rwLock = _locks.GetOrAdd(requestId, _ => new AsyncReaderWriterLock());
         return await rwLock.ReaderLockAsync();
     }
 
-    public async Task<IDisposable> EnterWriteAsync(int requestId)
+    public async Task<IDisposable> EnterWriteAsync(long requestId)
     {
         var rwLock = _locks.GetOrAdd(requestId, _ => new AsyncReaderWriterLock());
         return await rwLock.WriterLockAsync();
